Revive model picking refactoring
